backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
passlib[bcrypt]==1.7.4

# HTTP client for Ollama and ChromaDB
httpx==0.25.1
aiohttp==3.9.1

# OCR
easyocr==1.7.0
pytesseract==0.3.10
Pillow==10.1.0
opencv-python-headless==4.8.1.78
pdf2image==1.16.3

# Vector database
chromadb==0.4.18

# Data processing
pandas==2.1.3
numpy==1.26.2

# ML/Analysis (optional)
scikit-learn==1.3.2
xgboost==2.0.2

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
email-validator==2.1.0

# Utilities
python-dateutil==2.8.2

